Changed player's skills array to be uint32_t instead of uint8_t. Removed the special 254/255 values from the skill list so skills with those values actually show up properly in the list.
git-svn-id: https://svn.code.sf.net/p/seq/svn/showeq/branches/pre_5_0_beta@566 19a12193-7728-0410-8f54-e9f61ebc0c9f

+++ src/player.h
@@ -100,7 +100,7 @@ public:
    int16_t headingDegrees() const { return m_headingDegrees; }
    bool validPos() const { return m_validPos; }
 
-   uint8_t getSkill(uint8_t skillId) { return m_playerSkills[skillId]; }
+   uint32_t getSkill(uint8_t skillId) { return m_playerSkills[skillId]; }
    uint8_t getLanguage(uint8_t langId) { return m_playerLanguages[langId]; }
    
    int getPlusHP() { return m_plusHP; }
@@ -210,7 +210,7 @@ public:
   uint16_t m_defaultDeity;
   uint8_t m_defaultClass;
   uint8_t m_defaultLevel;
-  uint8_t m_playerSkills[MAX_KNOWN_SKILLS];
+  uint32_t m_playerSkills[MAX_KNOWN_SKILLS];
   uint8_t m_playerLanguages[MAX_KNOWN_LANGS];
   
   uint16_t m_plusMana;
